+++ nts.uk/uk.pr/pr.ctx/pr.core/nts.uk.ctx.pr.core.dom/src/main/java/nts/uk/ctx/pr/core/dom/wageprovision/formula/BasicCalculationFormula.java
@@ -51,7 +51,7 @@ public class BasicCalculationFormula extends AggregateRoot {
     */
     private Optional<BasicCalculationForm> basicCalculationForm;
     
-    public BasicCalculationFormula(String formulaCode, String historyID, String masterUseCode, int calculationFormulaClassification, BigDecimal basicCalculationFormula, BasicCalculationForm basicCalculationForm) {
+    public BasicCalculationFormula(String formulaCode, String historyID, String masterUseCode, int calculationFormulaClassification, Long basicCalculationFormula, BasicCalculationForm basicCalculationForm) {
         this.formulaCode = new FormulaCode(formulaCode);
         this.historyID = historyID;
         this.masterUseCode = new MasterUseCode(masterUseCode);
